pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "product-service"
version = "1.0.0"
description = "High-load microservice for product family management"
authors = ["Zashita LTD <dev@zashita.ltd>"]
readme = "README.md"
packages = [
    { include = "internal" },
    { include = "pkg" },
    { include = "cmd" },
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = {extras = ["standard"], version = "^0.27.0"}
asyncpg = "^0.29.0"
aiokafka = "^0.10.0"
redis = "^5.0.1"
msgpack = "^1.0.7"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
google-cloud-aiplatform = "^1.38.1"
circuitbreaker = "^2.0.0"
python-dotenv = "^1.0.0"
structlog = "^24.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.23.3"
pytest-cov = "^4.1.0"
httpx = "^0.26.0"
faker = "^22.2.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip = [".venv", "migrations"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
exclude = ["migrations", "tests"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=internal --cov=pkg --cov-report=term-missing"

[tool.coverage.run]
source = ["internal", "pkg"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
